+++ Controllers/DasKeyboardController/RGBController_DasKeyboard.cpp
@@ -247,7 +247,7 @@ RGBController_DasKeyboard::~RGBController_DasKeyboard()
     /*---------------------------------------------------------*\
     | Delete the matrix map                                     |
     \*---------------------------------------------------------*/
-    unsigned int zone_size = zones.size();
+    unsigned int zone_size = (unsigned int)zones.size();
 
     for(unsigned int zone_index = 0; zone_index < zone_size; zone_index++)
     {
